It is working

+++ sempy/types.py
@@ -1,3 +1,3 @@
 import numpy as np
 
-SEMPY_SCALAR=np.float32
+SEMPY_SCALAR=np.float64
